Extract geocoding logic
