+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/common/constants.kt
@@ -1,0 +1,6 @@
+package app.ehrenamtskarte.backend.common
+
+const val COUNTRY_CODE = "de"
+const val STATE = "Bayern"
+
+
