+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/steps/Sanitize.kt
@@ -1,20 +1,14 @@
 package app.ehrenamtskarte.backend.stores.importer.steps
 
+import app.ehrenamtskarte.backend.common.STATE
+import app.ehrenamtskarte.backend.stores.geocoding.queryFeatures
 import app.ehrenamtskarte.backend.stores.importer.PipelineStep
 import app.ehrenamtskarte.backend.stores.importer.types.AcceptingStore
-import com.fasterxml.jackson.databind.ObjectMapper
-import io.ktor.client.HttpClient
-import io.ktor.client.request.request
-import io.ktor.http.*
 import kotlinx.coroutines.runBlocking
 import org.geojson.Feature
-import org.geojson.FeatureCollection
 import org.geojson.Point
 import org.slf4j.Logger
-import java.io.File
 import kotlin.math.*
-
-const val STATE = "Bayern"
 
 const val DISTANCE_THRESHOLD_IN_KM = 5
 
@@ -24,7 +18,7 @@ const val STREET_EXCLUDE_PATTERN = "Postfach"
 class Sanitize(private val logger: Logger) : PipelineStep<List<AcceptingStore>, List<AcceptingStore>>() {
 
     override fun execute(input: List<AcceptingStore>): List<AcceptingStore> = runBlocking {
-        val stateFeature = query(listOf(Pair("state", STATE))).first()
+        val stateFeature = queryFeatures(listOf(Pair("state", STATE))).first()
         input.map { store ->
             if (store.street != null && store.street.contains(STREET_EXCLUDE_PATTERN)) return@map store
 
@@ -32,7 +26,7 @@ class Sanitize(private val logger: Logger) : PipelineStep<List<AcceptingStore>, 
             val postalCode = store.postalCode
 
             val postalCodeFeature = if (postalCode != null) {
-                query(listOf(Pair("postalcode", postalCode))).firstOrNull()
+                queryFeatures(listOf(Pair("postalcode", postalCode))).firstOrNull()
             } else null
 
             if (postalCodeFeature == null) {
@@ -102,59 +96,6 @@ class Sanitize(private val logger: Logger) : PipelineStep<List<AcceptingStore>, 
         return earthRadiusKm * c
     }
 
-    private suspend fun query(params: List<Pair<String, String>>): List<Feature> {
-        val fileName = params.joinToString("_") { it.second }
-            .replace(" ", "_")
-            .replace("/", "")
-            .replace(".", "")
-
-        val file = File("${System.getProperty("user.dir")}/data/nominatim/v1/$fileName.json")
-
-        val geoJson = if (file.exists()) {
-            // Use a cached version on disk.
-            file.readText()
-        } else {
-            val client = HttpClient()
-            val response = client.request<String> {
-                url {
-                    protocol = URLProtocol.HTTP
-                    host = "nominatim.maps.tuerantuer.org"
-                    path("nominatim", "search")
-                    parameters.append("format", "geojson")
-                    parameters.append("addressdetails", "1")
-                    parameters.append("countrycodes", COUNTRY_CODE)
-                    params.forEach {
-                        parameters.append(it.first, it.second)
-                    }
-                }
-                method = HttpMethod.Get
-            }
-            file.parentFile.mkdirs()
-            file.writeText(response)
-            response
-        }
-
-        return ObjectMapper().readValue<FeatureCollection>(geoJson, FeatureCollection::class.java).features
-    }
-
-    private suspend fun queryFeatures(store: AcceptingStore): List<Feature> {
-        val street = store.street
-        val houseNumber = store.houseNumber
-
-        val baseParameters = listOf(Pair("state", STATE), Pair("city", store.location))
-
-        if (street != null && houseNumber != null) {
-            val address = listOf(houseNumber, street).joinToString(" ")
-            val features = query(baseParameters.plus(Pair("street", address)))
-            if (features.isNotEmpty()) return features
-        }
-        if (street != null) {
-            val features = query(baseParameters.plus(Pair("street", street)))
-            if (features.isNotEmpty()) return features
-        }
-
-        return query(baseParameters)
-    }
 
     private fun Feature.latitude(): Double = (geometry as Point).coordinates.latitude
     private fun Feature.longitude(): Double = (geometry as Point).coordinates.longitude
